beautify examples
