+++ examples/templates/examples/example_frequency.html
@@ -1,5 +1,10 @@
 {% extends "base.html" %}
 {% block main %}
+    <a class="btn btn-info" href="{% url 'examples:example_list' %}">
+              <span data-feather="file"></span>
+              Example list
+            </a>
+<p></p>
     <form method="post">
         {% csrf_token %}
         {{form.as_p}}
