+++ laboratory/urls.py
@@ -6,6 +6,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
+    path('examples/', include('examples.urls')),
     path('', include('core.urls')),
     path('analysis/', include('analysis.urls')),
     path("__debug__/", include("debug_toolbar.urls")),
